pyproject.toml
[project]
name = "livekit-tavus-backend-skeleton"
version = "0.1.0"
description = "Backend skeleton for LiveKit + Tavus prototype"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn",
  "python-dotenv",
  "livekit-api",
  "livekit-agents",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "agent"]
